+++ docs/changelog.rst
@@ -1,6 +1,8 @@
 =========
 Changelog
 =========
+
+* :bug:`1491` All aave historical events should now be properly returned. Not only interest events.
 * :bug:`1482` Use binance api server time to avoid clock skew error with the signatures
 * :feature:`-` Users can now easily copy the address from the blockchain account view.
 * :bug:`1453` Users will now see an validation error message when attempting to add an existing account.